Refactor code, use async/await
