+++ src/config/swaggerconfig.js
@@ -2,15 +2,15 @@ const swaggerJSDoc = require('swagger-jsdoc');
 
 const options = {
   definition: {
-    name: "Rides API",
+    name: 'Rides API',
     info: {
-      title: 'Rides API', 
+      title: 'Rides API',
       description: 'This is an API for creating and fetching rides. \n \n To make changes in the documentation, just update content found under @swagger decorator found in each route. \n \n For more information on syntax, visit: https://swagger.io/docs/specification/basic-structure/',
-      version: '1.0.0' 
+      version: '1.0.0',
     },
   },
   // Path to the API docs
-  apis: ['../controllers/*.js'],
+  apis: ['src/controllers/**/*.js'],
 };
 
-module.exports = swaggerJSDoc(options)
+module.exports = swaggerJSDoc(options);
