+++ src/config/winstonconfig.js
@@ -1,0 +1,21 @@
+const winston = require('winston');
+
+const expressWinston = require('express-winston');
+
+const infoLogger = expressWinston.logger({
+  transports: [
+    new winston.transports.Console(),
+  ],
+});
+
+const errorLogger = expressWinston.errorLogger({
+  transports: [
+    new winston.transports.Console(),
+    new winston.transports.File({ filename: 'error.log' }),
+  ],
+});
+
+module.exports = {
+  infoLogger,
+  errorLogger,
+};
